impl: StoreManager

+++ venus-worker/src/sealing/store/mod.rs
@@ -1,12 +1,16 @@
-use std::fs::{self, create_dir_all, read, remove_dir_all};
+use std::collections::BTreeMap;
+use std::fs::{self, create_dir_all, read, read_dir, remove_dir_all};
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::time::Duration;
 
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use serde::{Deserialize, Serialize};
+use tracing::warn;
 
 use crate::metadb::rocks::RocksMeta;
+
+mod util;
 
 const SUB_PATH_DATA: &str = "data";
 const SUB_PATH_META: &str = "meta";
@@ -45,10 +49,43 @@ impl Default for Config {
     }
 }
 
+/// storage location
+#[derive(Debug)]
+pub struct Location(PathBuf);
+
+impl AsRef<Path> for Location {
+    fn as_ref(&self) -> &Path {
+        &self.0
+    }
+}
+
+impl Location {
+    fn meta_path(&self) -> PathBuf {
+        self.0.join(SUB_PATH_META)
+    }
+
+    fn data_path(&self) -> PathBuf {
+        self.0.join(SUB_PATH_DATA)
+    }
+
+    fn config_file(&self) -> PathBuf {
+        self.0.join(CONFIG_FILE_NAME)
+    }
+}
+
+/// storage used in bytes
+pub struct Usage {
+    /// storage used for sealing data
+    pub data: u64,
+
+    /// storage used for metadb
+    pub meta: u64,
+}
+
 /// a single sealing store
 pub struct Store {
     /// storage location
-    pub location: PathBuf,
+    pub location: Location,
 
     /// sub path for data dir
     pub data_path: PathBuf,
@@ -58,14 +95,16 @@ pub struct Store {
 
     /// embedded meta database for current store
     pub meta: RocksMeta,
+    meta_path: PathBuf,
 }
 
 impl Store {
     /// initialize the store at given location
     pub fn init<P: AsRef<Path>>(loc: P, capacity: u64) -> Result<Self> {
-        create_dir_all(loc.as_ref())?;
+        let location = Location(loc.as_ref().to_owned());
+        create_dir_all(&location)?;
 
-        let data_path = loc.as_ref().join(SUB_PATH_DATA);
+        let data_path = location.data_path();
         create_dir_all(&data_path)?;
 
         let mut config = Config::default();
@@ -75,37 +114,85 @@ impl Store {
         fs::OpenOptions::new()
             .write(true)
             .create_new(true)
-            .open(loc.as_ref().join(CONFIG_FILE_NAME))
+            .open(location.config_file())
             .and_then(|mut f| f.write_all(&cfg_content))?;
 
-        let meta = RocksMeta::open(loc.as_ref().join(SUB_PATH_META))?;
+        let meta_path = location.meta_path();
+        let meta = RocksMeta::open(&meta_path)?;
 
         Ok(Store {
-            location: loc.as_ref().to_owned(),
+            location,
             data_path,
             config,
             meta,
+            meta_path,
         })
     }
 
     /// opens the store at given location
     pub fn open<P: AsRef<Path>>(loc: P) -> Result<Self> {
-        let content = read(loc.as_ref().join(CONFIG_FILE_NAME))?;
+        let location = Location(loc.as_ref().to_owned());
+        let data_path = location.data_path();
+        if !data_path.symlink_metadata()?.is_dir() {
+            return Err(anyhow!("{:?} is not a dir", data_path));
+        }
+
+        let content = read(location.config_file())?;
         let config: Config = toml::from_slice(&content)?;
 
-        let meta = RocksMeta::open(loc.as_ref().join(SUB_PATH_META))?;
+        let meta_path = location.meta_path();
+        let meta = RocksMeta::open(&meta_path)?;
 
         Ok(Store {
-            location: loc.as_ref().to_owned(),
-            data_path: loc.as_ref().join(SUB_PATH_DATA),
+            location,
+            data_path,
             config,
             meta,
+            meta_path,
+        })
+    }
+
+    /// returns the disk usages inside the store
+    pub fn usage(&self) -> Result<Usage> {
+        let meta_used = util::disk_usage(&self.meta_path)?;
+        let data_used = util::disk_usage(&self.data_path)?;
+        Ok(Usage {
+            meta: meta_used,
+            data: data_used,
         })
     }
 
     /// cleanup cleans the store
     pub fn cleanup(&self) -> Result<()> {
-        remove_dir_all(&self.data_path)?;
+        let entries: Vec<_> = read_dir(&self.data_path)?.collect();
+        if !entries.is_empty() {
+            remove_dir_all(&self.data_path)?;
+            create_dir_all(&self.data_path)?;
+        }
+        Ok(())
+    }
+}
+
+/// manages the sealing stores
+#[derive(Default)]
+pub struct StoreManager {
+    stores: BTreeMap<String, Store>,
+}
+
+impl StoreManager {
+    /// loads specific
+    pub fn load(list: Vec<String>) -> Result<()> {
+        let mut stores = BTreeMap::new();
+        for path in list {
+            if stores.get(&path).is_some() {
+                warn!(path = path.as_str(), "store already loaded");
+                continue;
+            }
+
+            let store = Store::open(Path::new(&path))?;
+            stores.insert(path, store);
+        }
+
         Ok(())
     }
 }
